requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-stubs-ext==5.2.2
djangorestframework==3.16.0
idna==3.10
requests==2.32.4
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
